Add tests for S-S dimer chemistry
The explicit chemistry that could be occuring had not been automated
in a test before. These new tests should check to make sure that we
don't run into similar hashing problems in the future, by confirming
the correct chemistry for the simplest of cases.

To see the effect of hash collisions, change the hasing method used
to one of the other options, and there is no longer the ability to
distinguish between events with the same bond type and partners,
regardless of order. This is very problematic for all of the bi-
molecular reaction events, and should be addressed through eliminating
the use of explicit hashing.