Fix error involving hypervalent carbons
With some larger molecules, there was the chance that one of the
carbons would have a valency of 5, which would result in an error
within rdkit when trying to build the molecule. This bug should
now be squashed, as confirmed by the passing "testCheckForValenceError"
using pytest. Merging with master should close #8.